add uname -r support
