+++ kippo/commands/base.py
@@ -82,16 +82,6 @@ class command_hostname(HoneyPotCommand):
         self.writeln(self.honeypot.hostname)
 commands['/bin/hostname'] = command_hostname
 
-class command_uname(HoneyPotCommand):
-    def call(self):
-        if len(self.args) and self.args[0].strip() in ('-a', '--all'):
-            self.writeln(
-                'Linux %s 2.6.26-2-686 #1 SMP Wed Nov 4 20:45:37 UTC 2009 i686 GNU/Linux' % \
-                self.honeypot.hostname)
-        else:
-            self.writeln('Linux')
-commands['/bin/uname'] = command_uname
-
 class command_ps(HoneyPotCommand):
     def call(self):
         user = self.honeypot.user.username
